Conflicts:
	Delectable/app/src/main/AndroidManifest.xml
	Delectable/app/src/main/java/com/delectable/mobile/api/cache/UserInfo.java
	Delectable/app/src/main/java/com/delectable/mobile/ui/wineprofile/fragment/WineProfileFragment.java
	Delectable/app/src/main/java/com/delectable/mobile/ui/wineprofile/widget/ChooseVintageDialogRow.java
	Delectable/app/src/main/res/values/styles.xml

+++ Delectable/app/src/main/java/com/delectable/mobile/ui/common/widget/NumericRatingSeekBar.java
@@ -63,7 +63,7 @@ public class NumericRatingSeekBar extends RelativeLayout {
             @Override
             public void onRatingsChanged(int rating) {
 
-                float ratingOf10 = Rating.getRatingOfTenFrom40(rating);
+                double ratingOf10 = Rating.getRatingOfTenFrom40(rating);
                 String ratingStr = DECIMAL_FORMAT.format(ratingOf10);
                 mScoreTextView.setText(ratingStr);
